Move open_gz_file in archive_utils
